requirements.txt
dotenv
pydantic
pydantic_settings
torch
streamlit
pypdfium2
pdftext
regex
tqdm
filetype
ftfy
surya-ocr
bs4
rapidfuzz
markdown2
scikit-learn
markdownify==0.11.6
google-genai
pytesseract 
pillow
langchain==0.1.13
chromadb==0.4.24
pypdf==4.1.0
sentence_transformers
boto3
httpx==0.27.2
openai==1.55.3
python-dotenv
chromadb
fastapi 
uvicorn
ollama
requests
streamlit_option_menu
pymupdf
pandas
langchain-openai


sentence_transformers
boto3
openai
python-dotenv
pandas
chromadb
uvicorn
xlrd
streamlit
streamlit_option_menu
pysqlite3-binary